rename credit card package
